requirements.txt
django
djangorestframework
markdown
django-filter
coreapi
django-crispy-forms
django-guardian
django-oauth-toolkit